expose addMessage action (#17)
There's an issue in component lib where user message that result in an error will cause the user message to not be added to state. This is because, within the same useEffect call, the default error handler function uses `setMessages` to append the error message to the now **staled** state -- this happen because setNextMessage adds in the user message BUT the `messages` in React component fetch from `useChatState` doesn't get the updated state until the next render. This is an expected behavior from React. So, we need to expose a new action that appends new message to the current "in-progress" state.

Also update chat-core version to include the stream event bug fix.

J=CLIP-245
TEST=auto&manual

see that tests passed
replicate the scenario described above in the test-site and see that user's message is still rendered even though there's an error.