pass version to clientSdk field (#22)
extract chat-headless-react version to pass into `clientSdk` field in analytic event's payload

Due to the import of `package.json` the structure of `dist/` is reformatted (now all previous files are inside src/ and package.json is included in the respective esm and commonjs folder). As such, the entry points needs to be updated to reflect this.

J=CLIP-309
TEST=manual&auto

see that jest tests passed
see in test-site that clientSdk is passed into the payload with the correct versions (e.g. include the new CHAT_HEADLESS_REACT in addition to CHAT_HEADLESS and CHAT_CORE)

<img width="701" alt="Screenshot 2023-07-07 at 12 48 22 PM" src="https://github.com/yext/chat-headless/assets/36055303/4caefa1c-6f80-4fab-a54a-2eaff8968a6b">