ChatHeadless: upgrade @yext/analytics version dependency (#63)
Upgrades the version of @yext/analytics we use from 0.6.3 to 1.0.1.
This change requires some small tweaks to how we build the analytics
report service.

A change to note is that the new analytics SDK no longer supports
overriding the analytics endpoint and only supports sandbox and
production environments, so we cannot send analytics events when
testing with a dev chat bot in the test-site.

J=WAT-4671
TEST=auto, manual

Updated unit and integration tests.

Ran the test-site locally with a sandbox chat bot, was able to
interact with the bot normally and saw chat events being logged to
Snowflake. Also ran test-site with a dev chat bot and confirmed
that we can still interact with the bot, but the events API blocks
the event logging requests that get made.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>