ChatHeadless: update DeepPartial type
We want to allow ChatHeadless.report to be called with an event payload that only has a partially constructed chat field, because the report method will fill in those fields automatically from the config. However, the previous DeepPartial type from redux was not setting the chat field's nested fields to optional correctly, so it was throwing an error when report was called with a chat object missing the botId field. Defining a DeepPartial type ourselves resolves this.

Note that this does not change the DeepPartial type used by HeadlessConfig, only ChatHeadless.

J=WAT-4671
TEST=auto, manual

Tests pass. Spun up test site locally, saw analytics events going through to Events API as expected for responses and link clicks.